requirements.txt
streamlit
opencv-python-headless
torch
torchvision
ultralytics
numpy
Pillow